# openapi.yaml
openapi: 3.1.0
info:
  title: CoolProp Server
  description: REST API wrapping the complete C API of CoolProp v8.0.0 (thermophysical properties of pure fluids, mixtures and humid air). Every one of the 71 functions exported by CoolPropLib.h is exposed by exactly one endpoint — see the `coverage` module for the mapping.
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/fluids/add-json:
    post:
      tags:
      - routes::info
      summary: Add fluids defined by a JSON string to a backend's fluid library.
      description: Maps to the C function `add_fluids_as_JSON`.
      operationId: add_fluids_as_json
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFluidsRequest'
        required: true
      responses:
        '200':
          description: Fluids added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
        '400':
          description: CoolProp rejected the fluid definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/fluids/extract-backend:
    post:
      tags:
      - routes::info
      summary: Split a fluid string into its backend and fluid parts.
      description: Maps to the C function `C_extract_backend`.
      operationId: extract_backend
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractBackendRequest'
        required: true
      responses:
        '200':
          description: Backend and fluid parts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendFluid'
        '400':
          description: Extraction failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/fluids/is-valid:
    get:
      tags:
      - routes::info
      summary: Check whether a fluid string is valid.
      description: Maps to the C function `C_is_valid_fluid_string`.
      operationId: is_valid_fluid_string
      parameters:
      - name: name
        in: query
        description: Fluid string to validate, e.g. `"HEOS::Water[0.5]&Ethane[0.5]"`.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Validity flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlagValue'
  /api/v1/fluids/{fluid}/param/{param}:
    get:
      tags:
      - routes::info
      summary: 'Fluid metadata as a string: `"aliases"`, `"CAS"`, `"HASH"`, `"BibTeX"` ...'
      description: Maps to the C function `get_fluid_param_string`.
      operationId: get_fluid_param_string
      parameters:
      - name: fluid
        in: path
        description: Fluid name, e.g. `Water`
        required: true
        schema:
          type: string
      - name: param
        in: path
        description: Fluid parameter, e.g. `aliases`
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Fluid parameter value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringValue'
        '400':
          description: Unknown fluid/parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/fluids/{fluid}/param/{param}/length:
    get:
      tags:
      - routes::info
      summary: Length of the string that [`get_fluid_param_string`] returns.
      description: Maps to the C function `get_fluid_param_string_len`.
      operationId: get_fluid_param_string_len
      parameters:
      - name: fluid
        in: path
        description: Fluid name
        required: true
        schema:
          type: string
      - name: param
        in: path
        description: Fluid parameter
        required: true
        schema:
          type: string
      responses:
        '200':
          description: String length
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LengthValue'
        '400':
          description: Unknown fluid/parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/input-pairs/index:
    get:
      tags:
      - routes::info
      summary: Integer index for an AbstractState input pair name.
      description: Maps to the C function `get_input_pair_index`.
      operationId: get_input_pair_index
      parameters:
      - name: name
        in: query
        description: Parameter name, e.g. `"T"`, `"Dmolar"`, `"PT_INPUTS"`, ...
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Input pair index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexValue'
        '400':
          description: Unknown input pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/misc/debug-level:
    get:
      tags:
      - routes::info
      summary: Current CoolProp debug level.
      description: Maps to the C function `get_debug_level`.
      operationId: get_debug_level
      responses:
        '200':
          description: Debug level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexValue'
    put:
      tags:
      - routes::config
      summary: Set the CoolProp debug level (process-global).
      description: Maps to the C function `set_debug_level`.
      operationId: set_debug_level
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebugLevelRequest'
        required: true
      responses:
        '200':
          description: Debug level set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
  /api/v1/misc/f2k:
    get:
      tags:
      - routes::info
      summary: Convert degrees Fahrenheit to Kelvin.
      description: Maps to the C function `F2K`.
      operationId: f2k
      parameters:
      - name: t
        in: query
        description: Temperature.
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Temperature in K
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoubleValue'
  /api/v1/misc/k2f:
    get:
      tags:
      - routes::info
      summary: Convert Kelvin to degrees Fahrenheit.
      description: Maps to the C function `K2F`.
      operationId: k2f
      parameters:
      - name: t
        in: query
        description: Temperature.
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Temperature in °F
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoubleValue'
  /api/v1/misc/saturation-ancillary:
    get:
      tags:
      - routes::info
      summary: |-
        Evaluate a saturation ancillary curve (fast, but less accurate than the
        full EOS).
      description: Maps to the C function `saturation_ancillary`.
      operationId: saturation_ancillary
      parameters:
      - name: fluid
        in: query
        description: Fluid name (Helmholtz-EOS backend only).
        required: true
        schema:
          type: string
      - name: output
        in: query
        description: Desired output of the ancillary, e.g. `"P"`.
        required: true
        schema:
          type: string
      - name: q
        in: query
        description: 'Quality: 0 (saturated liquid) or 1 (saturated vapor).'
        required: true
        schema:
          type: integer
          format: int32
      - name: input
        in: query
        description: Ancillary input variable, e.g. `"T"`.
        required: true
        schema:
          type: string
      - name: value
        in: query
        description: Input value.
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Ancillary value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoubleValue'
        '400':
          description: CoolProp rejected the input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/params/global/{param}:
    get:
      tags:
      - routes::info
      summary: |-
        Query a global parameter string: `version`, `gitrevision`, `fluids_list`,
        `errstring`, `warnstring`, `FluidsList`, ...
      description: Maps to the C function `get_global_param_string`.
      operationId: get_global_param_string
      parameters:
      - name: param
        in: path
        description: Global parameter name, e.g. `version`
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Parameter value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringValue'
        '400':
          description: Unknown parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/params/index:
    get:
      tags:
      - routes::info
      summary: Integer index for a CoolProp parameter name (for keyed outputs).
      description: Maps to the C function `get_param_index`.
      operationId: get_param_index
      parameters:
      - name: name
        in: query
        description: Parameter name, e.g. `"T"`, `"Dmolar"`, `"PT_INPUTS"`, ...
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Parameter index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexValue'
        '400':
          description: Unknown parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/params/information/{param}:
    get:
      tags:
      - routes::info
      summary: |-
        Long description, units, or IO role of a CoolProp parameter, e.g. for
        `"T"` + `long` → `"Temperature [K]"`.
      description: Maps to the C function `get_parameter_information_string`.
      operationId: get_parameter_information_string
      parameters:
      - name: param
        in: path
        description: Parameter name, e.g. `Dmolar`
        required: true
        schema:
          type: string
      - name: info
        in: query
        description: |-
          Kind of information: `"long"` (description, default), `"short"`,
          `"units"`, or `"IO"`.
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: Human-readable parameter information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringValue'
        '400':
          description: Unknown parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
    Ack:
      type: object
      description: Simple acknowledgement for mutating calls.
      required:
      - success
      properties:
        success:
          type: boolean
    AddFluidsRequest:
      type: object
      required:
      - backend
      - fluid_string
      properties:
        backend:
          type: string
          description: Backend to add the fluids to, e.g. `"HEOS"`, `"SRK"`, `"PR"`.
        fluid_string:
          type: string
          description: JSON-formatted fluid definition string.
...
    BackendFluid:
      type: object
      required:
      - backend
      - fluid
      properties:
        backend:
          type: string
          description: Backend part (`"HEOS"`, `"REFPROP"`, ...; empty when not prefixed).
        fluid:
          type: string
          description: Fluid part, e.g. `"Water[0.5]&Ethane[0.5]"`.
...
    DebugLevelRequest:
      type: object
      required:
      - level
      properties:
        level:
          type: integer
          format: int32
          description: Verbosity level for CoolProp debugging output (0–10).
    DoubleValue:
      type: object
      description: Single scalar result.
      required:
      - value
      properties:
        value:
          type: number
          format: double
    ErrorDetail:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ExtractBackendRequest:
      type: object
      required:
      - fluid_string
      properties:
        fluid_string:
          type: string
          description: Fluid string, e.g. `"REFPROP::Water[0.5]&Ethane[0.5]"`.
...
    FlagValue:
      type: object
      description: Boolean flag result.
      required:
      - value
      properties:
        value:
          type: boolean
...
    IndexValue:
      type: object
      description: Single integer index result.
      required:
      - index
      properties:
        index:
          type: integer
          format: int64
...
    LengthValue:
      type: object
      description: Single length result.
      required:
      - length
      properties:
        length:
          type: integer
          format: int64
...
    StringValue:
      type: object
      description: Single string result.
      required:
      - value
      properties:
        value:
          type: string
...